requirement.txt
opencv-python
matplotlib
numpy
Pillow
torch
torchvision
tqdm
pyyaml
requests
pandas
seaborn
xlsxwriter
ultralytics
# tkinter (built-in with Python)